# contracts/backend/openapi.yaml
openapi: 3.1.0
info:
  title: DIP3 Java Analysis Backend
  version: 2.0.0
paths:
...
  /api/analysis/sessions:
    post:
      security: [{ cookieAuth: [] }]
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema: { type: object, required: [question], properties: { question: { type: string, maxLength: 300 } } }
          multipart/form-data:
            schema: { type: object, required: [question], properties: { question: { type: string, maxLength: 300 } } }
      responses:
        '303': { description: Created or login/error redirect, headers: { Location: { schema: { type: string } } } }
        '422': { $ref: '#/components/responses/Error' }
        '500': { description: Unexpected failure, content: { application/json: { schema: { $ref: './schemas/error.schema.json' } } } }
...
  /api/analysis/sessions/{sessionId}/execute:
    post:
      summary: Submit a root or follow-up execution through its isolated Java job contract
      security: [{ cookieAuth: [] }]
      parameters:
        - { name: sessionId, in: path, required: true, schema: { type: string, minLength: 1 } }
        - { name: Idempotency-Key, in: header, required: false, description: Missing or blank uses the stable initial-execution key for this session, schema: { type: string, maxLength: 128 } }
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema: { type: object, properties: { followUpId: { type: string, maxLength: 200, description: Blank or absent submits the root java-initial-v1 round; a non-blank ID submits java-follow-up-v1 } } }
          multipart/form-data:
            schema: { type: object, properties: { followUpId: { type: string, maxLength: 200, description: Blank or absent submits the root java-initial-v1 round; a non-blank ID submits java-follow-up-v1 } } }
      responses:
        '303': { description: Submitted or login redirect, headers: { Location: { schema: { type: string } } } }
        '400': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { description: Existing idempotency conflict, a legacy session, or a follow-up already bound to another execution, content: { application/json: { schema: { $ref: './schemas/error.schema.json' } } } }
        '422': { description: Follow-up context is not executable or must be replanned first; browser form requests receive a 303 redirect carrying followUpExecutionError, content: { application/json: { schema: { $ref: './schemas/error.schema.json' } } } }
        '500': { $ref: '#/components/responses/Error' }
...
  /api/analysis/sessions/{sessionId}/stream:
    get:
      security: [{ cookieAuth: [] }]
      parameters:
        - { name: sessionId, in: path, required: true, schema: { type: string, minLength: 1 } }
        - { name: executionId, in: query, required: true, schema: { type: string, minLength: 1 } }
        - { name: afterSequence, in: query, required: false, schema: { type: integer, minimum: 0, default: 0 } }
      responses:
        '200':
          description: SSE frames; each data payload is an ExecutionEvent
          headers:
            Cache-Control: { schema: { type: string } }
            X-Accel-Buffering: { schema: { type: string, const: no } }
          content:
            text/event-stream: { schema: { type: string } }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
  /internal/analysis/sessions/{sessionId}/executions/{executionId}/snapshot:
    get:
      security: [{ cookieAuth: [] }]
      parameters:
        - { name: sessionId, in: path, required: true, schema: { type: string } }
        - { name: executionId, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: PostgreSQL recovery fact, content: { application/json: { schema: { $ref: './schemas/execution-snapshot.schema.json' } } } }
        '401': { $ref: '#/components/responses/Error' }
        '404': { description: Snapshot not found }
...
components:
  securitySchemes:
    cookieAuth: { type: apiKey, in: cookie, name: dip3_session }
    graphSyncOpsSecret: { type: apiKey, in: header, name: X-Graph-Sync-Ops-Secret }
  responses:
    Error:
      description: Explicit failure
      content:
        application/json: { schema: { $ref: './schemas/error.schema.json' } }
...